Add files via upload
Voting Management System is a voting system which is concerned with online voting mechanism. The voting system is having mainly two modules which will be the foundation of this project. First module is at Administration Level and second module is of Employees. This is a java based project which will run in the java environment currently. It maintains authentication in order to access the application. Administrator task includes Employee Registration i.e., new employee recruitment will take place.  News and information about company can access employee.